requirements.txt
python-dotenv
langchain
langchain-chroma
langchain-community
langchain-openai
langchain-huggingface
beautifulsoup4
sentence-transformers